Move `Host` to dedicated directory
