+++ crates/fj-host/src/lib.rs
@@ -15,6 +15,7 @@
 
 #![warn(missing_docs)]
 
+mod host;
 mod parameters;
 mod platform;
 mod watcher;
@@ -30,6 +31,7 @@ use std::{
 
 use fj::{abi, version::RawVersion};
 use fj_interop::status_report::StatusReport;
+use host::Host;
 use thiserror::Error;
 
 use self::platform::HostPlatform;
@@ -288,34 +290,4 @@ pub enum Error {
         /// workspace root.
         possible_paths: Vec<PathBuf>,
     },
-}
-
-struct Host<'a> {
-    args: &'a Parameters,
-    model: Option<Box<dyn fj::models::Model>>,
-}
-
-impl<'a> Host<'a> {
-    pub fn new(parameters: &'a Parameters) -> Self {
-        Self {
-            args: parameters,
-            model: None,
-        }
-    }
-
-    pub fn take_model(&mut self) -> Option<Box<dyn fj::models::Model>> {
-        self.model.take()
-    }
-}
-
-impl<'a> fj::models::Host for Host<'a> {
-    fn register_boxed_model(&mut self, model: Box<dyn fj::models::Model>) {
-        self.model = Some(model);
-    }
-}
-
-impl<'a> fj::models::Context for Host<'a> {
-    fn get_argument(&self, name: &str) -> Option<&str> {
-        self.args.get(name).map(|s| s.as_str())
-    }
 }
